refactor trends
