+++ src/commands/trends/kdtrend.js
@@ -1,0 +1,7 @@
+import SingleTrendCommand from "./singleTrendCommand";
+
+export default class KDTrendCommand extends SingleTrendCommand {
+  constructor() {
+    super("charts Kill/Death Rate by season", "kd", "KD", "Kill/Death Rate");
+  }
+}
